requirements.txt
groq
streamlit
langchain
langchain_community
PyPDF
langchain-openai
langchain-groq
python-dotenv
langchain-google-genai